Added base class for model viewsets that use django-filters, and did some restructuring
