+++ govapp/apps/main/views.py
@@ -1,64 +1,38 @@
-from rest_framework import response, viewsets
-from rest_framework.decorators import action
+from django.db.models.query import QuerySet
+from django_filters import rest_framework as filters
+from rest_framework import viewsets
+from rest_framework_datatables.django_filters.backends import DatatablesFilterBackend
+
+from govapp.apps.main.mixins import KeyValueListMixin
 
 from .models import District, Region
 from .serializers import DistrictSerializer, RegionSerializer
 
 
-class KeyValueListMixin:
-    @action(detail=False, methods=["get"], url_path="key-value-list")
-    def key_value_list(self, request):
-        if hasattr(self, "key_value_display_field"):
-            key_value_display_field = self.key_value_display_field
-        else:
-            class_model = self.get_serializer().Meta.model
-            field_names = [field.name for field in class_model._meta.get_fields()]
-            key_value_display_field = (
-                "display_name"
-                if "display_name" in field_names
-                else "name"
-                if "name" in field_names
-                else None
-            )
-            if not key_value_display_field:
+class DjangoFiltersModelViewSet(viewsets.ModelViewSet):
+    """Base class for viewsets that use DjangoFilters
+    Uses django-filter's filterset_class attribute if the action is in the list of actions.
+    Otherwise uses the default filter_queryset method.
+    """
+
+    filter_backends = [DatatablesFilterBackend, filters.DjangoFilterBackend]
+    actions = ["list"]  # The actions that will use the django_filters filterset_class
+
+    @property
+    def filterset_class(self):
+        if self.action in self.actions:
+            if not hasattr(self, "django_filters_filterset_class"):
                 raise AttributeError(
-                    "key_value_display_field is not defined on viewset or model"
+                    f"Expected {self.__class__.__name__} to have a django_filters_filterset_class attribute"
                 )
+            return self.django_filters_filterset_class
+        return None
 
-        queryset = self.get_queryset().only("id", key_value_display_field)
-        search_term = request.GET.get("term", "")
-        if search_term:
-            queryset = queryset.filter(
-                **{f"{key_value_display_field}__icontains": search_term}
-            )[:30]
-        serializer_class = self.get_key_value_serializer_class()
-        serializer = serializer_class(queryset, many=True)
-        return response.Response(serializer.data)
-
-    def get_key_value_serializer_class(self):
-        if hasattr(self, "key_value_serializer_class"):
-            return self.key_value_serializer_class
-
-        class_model = self.get_serializer().Meta.model
-        from rest_framework import serializers
-
-        class NamedModelKeyValueSerializer(serializers.ModelSerializer):
-            key = serializers.IntegerField(source="id")
-            value = serializers.SerializerMethodField()
-
-            class Meta:
-                model = class_model
-                fields = ["key", "value"]
-
-            def get_value(self, obj):
-                display_name = getattr(obj, "display_name", None)
-                if not display_name:
-                    return getattr(
-                        obj, "name", "No display_name or name field found on model"
-                    )
-                return display_name
-
-        return NamedModelKeyValueSerializer
+    def filter_queryset(self, queryset: QuerySet) -> QuerySet:
+        # If the action is in the list of actions then use the filterset_class
+        if self.filterset_class:
+            return self.filterset_class(self.request.GET, queryset=self.queryset).qs
+        return super().filter_queryset(queryset)
 
 
 class RegionViewSet(KeyValueListMixin, viewsets.GenericViewSet):
